backend/requirements.txt
# POD Multi-Agent System Dependencies
# =====================================

# Core LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
openai>=1.40.0
anthropic>=0.34.0

# HTTP Client
httpx>=0.27.0
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Database (for Checkpoint persistence)
psycopg2-binary>=2.9.9
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.8.0

# Web Framework (FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.0.0
python-multipart>=0.0.9

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0

# Optional: LangSmith for debugging
# langsmith>=0.1.0

# Langfuse for LLM observability
langfuse>=2.0.0

# Pillow for local mockup generation
Pillow>=10.0.0
